Изменил структуру каталогов в проекте
