+++ MetricsManager/MetricsAgent/DAL/Interfaces/IRepository.cs
