+++ MetricsManager/MetricsAgent/DAL/Models/CpuMetric.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.Models
+namespace MetricsAgent.DAL
 {
-	public class DotNetMetric
+	public class CpuMetric
 	{
 		public int Id { get; set; }
