+++ MetricsManager/MetricsAgent/DAL/Models/DotNetMetric.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.Models
+namespace MetricsAgent.DAL
 {
-	public class CpuMetric
+	public class DotNetMetric
 	{
 		public int Id { get; set; }
