+++ MetricsManager/MetricsAgent/DAL/Models/HddMetric.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.Models
+namespace MetricsAgent.DAL
 {
-	public class RamMetric
+	public class HddMetric
 	{
 		public int Id { get; set; }
