+++ MetricsManager/MetricsAgent/DAL/Models/NetworkMetric.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.Models
+namespace MetricsAgent.DAL
 {
-	public class HddMetric
+	public class NetworkMetric
 	{
 		public int Id { get; set; }
