+++ MetricsManager/MetricsAgent/DAL/Models/RamMetric.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.Models
+namespace MetricsAgent.DAL
 {
-	public class NetworkMetric
+	public class RamMetric
 	{
 		public int Id { get; set; }
